split handlers by files
